[feature/View]feat: Implement ViewModelTestView
